Add teams to trainings

+++ src/rest/Trainings/Actions/TrainingUpdateAction.php
@@ -44,6 +44,7 @@ class TrainingUpdateAction
                 'contain' => [
                     'TrainingDefinition',
                     'TrainingCoaches',
+                    'Teams',
                     'Season',
                     'Event',
                     'Event.Contents'
@@ -81,6 +82,15 @@ class TrainingUpdateAction
                 $table->TrainingCoaches,
                 false
             ))->validate($data);
+
+            $teamData = \JmesPath\search('data.relationships.teams', $data);
+            if (isset($teamData)) {
+                $teams = (new EntityExistValidator(
+                    'data.relationships.teams',
+                    $table->Teams,
+                    false
+                ))->validate($data);
+            }
 
             $defData = \JmesPath\search('data.relationships.definition.data', $data);
             if (isset($defData)) {
@@ -130,6 +140,10 @@ class TrainingUpdateAction
             $training->dirty('event', true);
             $training->event->user = $request->getAttribute('clubman.user');
 
+            if (isset($teams)) {
+                $training->teams = $teams;
+            }
+
             (new \REST\Trainings\TrainingValidator())->validate($training);
 
             $table->save($training, [
@@ -139,37 +153,38 @@ class TrainingUpdateAction
             ]);
 
             // Update coaches
-
-            // When a coach is not passed to this function, it must be deleted
-            $lookup = array_column($coaches, null, 'id');
-            $toDelete = [];
-            foreach ($training->coaches as $coach) {
-                if (!$lookup[$coach->id]) {
-                    $toDelete[] = $coach;
+            if ($coaches) {
+                // When a coach is not passed to this function, it must be deleted
+                $lookup = array_column($coaches, null, 'id');
+                $toDelete = [];
+                foreach ($training->coaches as $coach) {
+                    if (!$lookup[$coach->id]) {
+                        $toDelete[] = $coach;
+                    }
                 }
-            }
-            if (count($toDelete) > 0) {
-                $table->TrainingCoaches->unlink($training, $toDelete);
-            }
-
-            // When a coach is passed to this function and it's not in the
-            // table, it must be insert
-            $lookup = array_column($training->coaches, null, 'id');
-            $toInsert = [];
-            foreach ($coaches as $coach) {
-                if (!$lookup[$coach->id]) {
-                    $coach->_joinData = new Entity([
-                        'coach_type' => 0,
-                        'present' => false,
-                        'user' => $request->getAttribute('clubman.user')
-                    ], [
-                        'markNew' => true
-                    ]);
-                    $toInsert[] = $coach;
+                if (count($toDelete) > 0) {
+                    $table->TrainingCoaches->unlink($training, $toDelete);
                 }
-            }
-            if (count($toInsert) > 0) {
-                $table->TrainingCoaches->link($training, $toInsert);
+
+                // When a coach is passed to this function and it's not in the
+                // table, it must be insert
+                $lookup = array_column($training->coaches, null, 'id');
+                $toInsert = [];
+                foreach ($coaches as $coach) {
+                    if (!$lookup[$coach->id]) {
+                        $coach->_joinData = new Entity([
+                            'coach_type' => 0,
+                            'present' => false,
+                            'user' => $request->getAttribute('clubman.user')
+                        ], [
+                            'markNew' => true
+                        ]);
+                        $toInsert[] = $coach;
+                    }
+                }
+                if (count($toInsert) > 0) {
+                    $table->TrainingCoaches->link($training, $toInsert);
+                }
             }
 
             $route = $request->getAttribute('route');
